IR phantom check (#260)
* boilerplate

* gather events

* more things

* invoke

* need to fix

* add interface info for all

* non-fails passing now

* add phantom tests

* remove option

* remove comments

* forgot some

* more comments oops

* change back to old interfaces

* remove unused fn

* remove unused field

* misc

* change to old funcs

* remove default

* comment

* undo unneeded changes

* restructure to allow multiple errors and loop check

* update tests

* test reporting multiple phantom errors

---------

+++ src/ir_passes/discharge.rs
@@ -58,7 +58,7 @@ pub struct Discharge {
     // Diagnostics to be reported
     diagnostics: Vec<cr::Diagnostic<usize>>,
     /// Number of errors encountered
-    error_count: u32,
+    error_count: u64,
 }
 
 impl Construct for Discharge {
@@ -430,7 +430,7 @@ impl Visitor for Discharge {
         }
     }
 
-    fn after_traversal(&mut self) -> Option<u32> {
+    fn after_traversal(&mut self) -> Option<u64> {
         if self.error_count > 0 {
             Some(self.error_count)
         } else {
